Agregar mapstruct-processor al pom y corregir tipo Long en cadena Auditoria ._.

+++ src/main/java/com/uisrael/asistencia/aplicacion/casosuso/entrada/IAuditoriaUseCase.java
@@ -9,10 +9,10 @@ public interface IAuditoriaUseCase {
 	
 	Auditoria guardar(Auditoria nuevaAuditoria);
 
-	Auditoria buscarPorId(int idAuditoria);
+	Auditoria buscarPorId(Long idAuditoria);
 
 	List<Auditoria> listarTodos();
 
-	void eliminar(int idAuditoria);
+	void eliminar(Long idAuditoria);
 
 }
